renamed AllPodcastsContainer to DisplayPodcastsCo tainer (will be also used for subscribed page)
